+++ programs/ekr_loop.c
@@ -139,7 +139,7 @@ receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
 			       rcv.rcv_sid,
 			       rcv.rcv_ssn,
 			       rcv.rcv_tsn,
-			       ntohl(rcv.rcv_ppid),
+			       (unsigned int)ntohl(rcv.rcv_ppid),
 			       rcv.rcv_context,
 			       flags);
 		}
