+++ programs/ekr_loop_offload.c
@@ -96,7 +96,7 @@ handle_packets(void *arg)
 					usrsctp_conninput(fdp, buffer, (size_t)length, 0);
 				} else {
 					fprintf(stderr, "Wrong CRC32c: expected %08x received %08x\n",
-					        ntohl(computed_crc32c), ntohl(received_crc32c));
+					        (unsigned int)ntohl(computed_crc32c), (unsigned int)ntohl(received_crc32c));
 				}
 			} else {
 				fprintf(stderr, "Packet too short: length %zu", (size_t)length);
@@ -160,7 +160,7 @@ receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
 			       rcv.rcv_sid,
 			       rcv.rcv_ssn,
 			       rcv.rcv_tsn,
-			       ntohl(rcv.rcv_ppid),
+			       (unsigned int)ntohl(rcv.rcv_ppid),
 			       rcv.rcv_context,
 			       flags);
 		}
