+++ programs/ekr_server.c
@@ -140,7 +140,7 @@ receive_cb(struct socket *s, union sctp_sockstore addr, void *data,
 			       rcv.rcv_sid,
 			       rcv.rcv_ssn,
 			       rcv.rcv_tsn,
-			       ntohl(rcv.rcv_ppid),
+			       (unsigned int)ntohl(rcv.rcv_ppid),
 			       rcv.rcv_context);
 		}
 		free(data);
